+++ requirements.txt
@@ -1,0 +1,2 @@
+numpy~=1.26.4
+pandas>=2.1.4
